% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoiser_processing.R
\name{aad_calc}
\alias{aad_calc}
\title{Calculate Average Absolute Deviation}
\usage{
aad_calc(x, center = mode_function(x))
}
\arguments{
\item{x}{Numeric vector}

\item{center}{Numeric; center value (default is mode)}
}
\value{
Numeric; average absolute deviation
}
\description{
Helper function to calculate AAD from a specified center
}
\keyword{internal}
