% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_visualization_plots.R
\name{calculate_label_cex}
\alias{calculate_label_cex}
\title{Calculate optimal label cex for dendrogram}
\usage{
calculate_label_cex(n_samples)
}
\arguments{
\item{n_samples}{Number of samples in the dendrogram}
}
\value{
Numeric cex value between 0. 3 and 1.0
}
\description{
Calculate optimal label cex for dendrogram
}
