% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pn_limma.R
\name{clean_factor_levels}
\alias{clean_factor_levels}
\title{Make Factor Levels Syntactically Valid}
\usage{
clean_factor_levels(metadata, variable)
}
\arguments{
\item{metadata}{Data frame}

\item{variable}{Character, variable name}
}
\value{
Data frame with cleaned variable
}
\description{
Make Factor Levels Syntactically Valid
}
\keyword{internal}
