% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_utils.R
\name{compare_denoise_parameters}
\alias{compare_denoise_parameters}
\title{Compare Denoising Parameters}
\usage{
compare_denoise_parameters(
  cutpoint_results,
  top_n = 10,
  sort_by = "TP_FP_ratio"
)
}
\arguments{
\item{cutpoint_results}{Data frame; combined results from multiple PC levels}

\item{top_n}{Integer; number of top results to return (default = 10)}

\item{sort_by}{Character; column to sort by (default = "TP_FP_ratio")}
}
\value{
Data frame of top parameter combinations
}
\description{
Create comparison table of different denoising parameter combinations
}
