% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot_functions.R
\name{create_density_plot}
\alias{create_density_plot}
\title{Create Density Plot (Core Function)}
\usage{
create_density_plot(
  density_plot_data,
  density_vline_data = NULL,
  sample_flag_data = NULL,
  flag_ylim = 2,
  col_num = 3,
  xlim = c(0, 16),
  facet = "Sample",
  scale = "auto"
)
}
\description{
Create Density Plot (Core Function)
}
