% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_functions.R
\name{darrow}
\alias{darrow}
\title{Debug Arrow Message}
\usage{
darrow(..., debug = FALSE, level = 0, id = NULL)
}
\arguments{
\item{...}{Message components}

\item{debug}{Enable debug output}

\item{level}{Indentation level}

\item{id}{Module ID (optional)}
}
\description{
Debug Arrow Message
}
