% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_functions.R
\name{dend}
\alias{dend}
\title{Debug Section End}
\usage{
dend(title, id = NULL, debug = FALSE)
}
\arguments{
\item{title}{Section title}

\item{id}{Module ID (optional)}

\item{debug}{Enable debug output}
}
\description{
Debug Section End
}
