% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_functions.R
\name{dstart}
\alias{dstart}
\title{Debug Section Start}
\usage{
dstart(title, debug = FALSE)
}
\arguments{
\item{title}{Section title}

\item{debug}{Enable debug output}
}
\description{
Debug Section Start
}
