% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_functions.R
\name{dwarn}
\alias{dwarn}
\title{Debug Warning Message}
\usage{
dwarn(..., debug = FALSE, level = 0)
}
\arguments{
\item{...}{Message components}

\item{debug}{Enable debug output}

\item{level}{Indentation level}
}
\description{
Debug Warning Message
}
