% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_add_fData_function}
\alias{ExpSet_add_fData_function}
\title{Add Columns to featureData (fData) of an ExpressionSet}
\usage{
ExpSet_add_fData_function(
  ExpSet,
  meta_add,
  add_cols,
  feature_column = "Protein"
)
}
\arguments{
\item{ExpSet}{An \code{ExpressionSet} object.}

\item{meta_add}{A data frame with \code{Protein} and columns to add.}

\item{add_cols}{Character vector of column names from \code{meta_add} to add to \code{fData}.}
}
\value{
The updated \code{ExpressionSet} object with modified \code{fData}.
}
\description{
Merges new feature metadata columns into the \code{fData} slot of an \code{ExpressionSet} object.
}
\note{
Version 1.0 from
ExpSet_functions.R
}
