% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_add_pData_function}
\alias{ExpSet_add_pData_function}
\title{Add Columns to phenoData (pData) of an ExpressionSet}
\usage{
ExpSet_add_pData_function(ExpSet, meta_add, add_cols)
}
\arguments{
\item{ExpSet}{An \code{ExpressionSet} object.}

\item{meta_add}{A data frame with \code{Sample} and columns to add.}

\item{add_cols}{Character vector of column names from \code{meta_add} to add to \code{pData}.}
}
\value{
The updated \code{ExpressionSet} object with modified \code{pData}.
}
\description{
Merges new metadata columns into the \code{pData} slot of an \code{ExpressionSet} object.
}
\note{
Version 1.0 from
ExpSet_functions.R
}
