% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_build_other_metadata}
\alias{ExpSet_build_other_metadata}
\title{Build Other Metadata Structure}
\usage{
ExpSet_build_other_metadata(dat, datCollate, QC = NULL, project_info = NULL)
}
\arguments{
\item{dat}{Data preprocessing list containing param and other elements}

\item{datCollate}{Data collation list containing param, QC, and other elements}

\item{QC}{QC list containing param, BGFG_stats, Avg_QC, and other elements}

\item{project_info}{Project information to include in metadata}
}
\value{
A list containing organized metadata including parameters, QC data, and manifest
}
\description{
Creates the "other" metadata structure for ExpressionSet experimentData.
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
other <- ExpSet_build_other_metadata(dat, datCollate, QC, project_information())
}
}
