% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_create_clinical}
\alias{ExpSet_create_clinical}
\title{Create Clinical ExpressionSet}
\usage{
ExpSet_create_clinical(datCollate, metadata, other, PSA_list = NULL)
}
\arguments{
\item{datCollate}{A list containing data collation results}

\item{metadata}{An AnnotatedDataFrame with sample metadata}

\item{other}{A list containing additional metadata to store in experimentData}

\item{PSA_list}{Optional character vector of proteins in PSA list}
}
\value{
An ExpressionSet object with clinical data assays
}
\description{
Creates an ExpressionSet containing clinical-ready normalized intensity values with metadata, CV and CV_flagged.
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
expset <- ExpSet_create_clinical(datCollate, metadata, other)
}
}
