% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_create_feature_data}
\alias{ExpSet_create_feature_data}
\title{Create Feature Data for ExpressionSet}
\usage{
ExpSet_create_feature_data(
  data,
  feature_df,
  PSA_list = NULL,
  feature_col = "Protein"
)
}
\arguments{
\item{data}{A data frame containing feature information}

\item{feature_df}{A data frame with additional feature annotations}

\item{PSA_list}{Optional character vector of proteins in PSA list}

\item{feature_col}{Character. Name of the feature column (default: "Protein")}
}
\value{
An AnnotatedDataFrame object with feature metadata
}
\description{
Creates an AnnotatedDataFrame for feature data with optional PSA list annotation.
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
data <- data.frame(Protein = c("P1", "P2"), data = c("A", "B"))
feature_df <- data.frame(Protein = c("P1", "P2"), Gene = c("G1", "G2"))
features <- ExpSet_create_feature_data(data, feature_df)
}
}
