% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_create_list}
\alias{ExpSet_create_list}
\title{Create Complete ExpressionSet List}
\usage{
ExpSet_create_list(
  dat,
  datCollate,
  QC = NULL,
  project_info = NULL,
  PSA_list = NULL
)
}
\arguments{
\item{dat}{Data preprocessing list containing param and other elements}

\item{datCollate}{Data collation list containing data, output, param, and QC elements}

\item{QC}{QC list containing param, BGFG_stats, Avg_QC (optional)}

\item{project_info}{Project information to include in metadata (optional)}

\item{PSA_list}{Optional character vector of proteins in PSA list}
}
\value{
A named list of ExpressionSet objects
}
\description{
Creates a complete list of ExpressionSet objects from data collation results.
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
expset_list <- ExpSet_create_list(dat, datCollate, QC, project_information())
}
}
