% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_create_mean}
\alias{ExpSet_create_mean}
\title{Create Mean ExpressionSet}
\usage{
ExpSet_create_mean(datCollate, metadata, params, PSA_list = NULL)
}
\arguments{
\item{datCollate}{A list containing data collation results}

\item{metadata}{An AnnotatedDataFrame with sample metadata}

\item{params}{A list of parameters used in preprocessing}

\item{PSA_list}{Optional character vector of proteins in PSA list}
}
\value{
An ExpressionSet object with mean data assays
}
\description{
Creates an ExpressionSet containing mean intensity values with CV and CV_flagged.
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
expset <- ExpSet_create_mean(datCollate, metadata, params)
}


}
