% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_create_metadata}
\alias{ExpSet_create_metadata}
\title{Create Metadata for ExpressionSet}
\usage{
ExpSet_create_metadata(datCollate, manifest_full)
}
\arguments{
\item{datCollate}{A list containing data collation results}

\item{manifest_full}{A data frame with complete manifest including QC information}
}
\value{
An AnnotatedDataFrame object with sample metadata
}
\description{
Creates an AnnotatedDataFrame for phenotype data (metadata) with sample type classification.
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
metadata <- ExpSet_create_metadata(datCollate, manifest_full)
}
}
