% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_create_neti_o}
\alias{ExpSet_create_neti_o}
\title{Create NetI ExpressionSet}
\usage{
ExpSet_create_neti_o(
  datCollate,
  metadata,
  params,
  failed_samples,
  PSA_list = NULL,
  round_digits = 3
)
}
\arguments{
\item{datCollate}{A list containing data collation results}

\item{metadata}{An AnnotatedDataFrame with sample metadata}

\item{failed_samples}{Character vector of failed sample names to exclude}

\item{PSA_list}{Optional character vector of proteins in PSA list}

\item{round_digits}{Integer. Number of decimal places to round to (default: 3)}
}
\value{
An ExpressionSet object with NetI data for passed samples
}
\description{
Creates an ExpressionSet containing imputed log2 intensity values for passed samples only.
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
expset <- ExpSet_create_neti(datCollate, metadata, failed_samples)
}


}
