% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_create_norm}
\alias{ExpSet_create_norm}
\title{Create Normalized ExpressionSet}
\usage{
ExpSet_create_norm(datCollate, metadata, params, PSA_list = NULL)
}
\arguments{
\item{datCollate}{A list containing data collation results}

\item{metadata}{An AnnotatedDataFrame with sample metadata}

\item{params}{A list of parameters used in preprocessing}

\item{PSA_list}{Optional character vector of proteins in PSA list}
}
\value{
An ExpressionSet object with normalized data assays
}
\description{
Creates an ExpressionSet containing normalized intensity values with row-centered data, CV and CV_flagged.
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
expset <- ExpSet_create_norm(datCollate, metadata, params)
}
}
