% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_create_rawdata}
\alias{ExpSet_create_rawdata}
\title{Create Raw Data ExpressionSet}
\usage{
ExpSet_create_rawdata(datCollate, metadata, params)
}
\arguments{
\item{datCollate}{A list containing data collation results}

\item{metadata}{An AnnotatedDataFrame with sample metadata}

\item{params}{A list of parameters used in preprocessing}
}
\value{
An ExpressionSet object with raw data assays
}
\description{
Creates an ExpressionSet containing raw intensity data (BG, FG, NetI).
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
expset <- ExpSet_create_rawdata(datCollate, metadata, params)
}
}
