% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_load_from_file}
\alias{ExpSet_load_from_file}
\title{Load ExpressionSet from RDS File}
\usage{
ExpSet_load_from_file(file_path)
}
\arguments{
\item{file_path}{Character. Path to the RDS file}
}
\value{
An ExpressionSet object or list of ExpressionSet objects
}
\description{
Loads an ExpressionSet or list of ExpressionSets from an RDS file with validation.
}
\note{
Version 1.0 from ExpSet_functions.R
}
\examples{
\dontrun{
expset <- ExpSet_load_from_file("path/to/expset.rds")
}
}
