% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{ExpSet_select_function_2}
\alias{ExpSet_select_function_2}
\title{Select ExpressionSet and Set exprs()}
\usage{
ExpSet_select_function_2(ExpSet_list, name)
}
\arguments{
\item{ExpSet_list}{List; list of ExpressionSets}

\item{name}{Character; name of assay to use (e.g., "clinical_loess_normalised")}
}
\value{
ExpressionSet with exprs() set to selected assay
}
\description{
Helper function to extract ExpressionSet from list and set exprs() to selected assay
}
