% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_func_dev_functions.R
\name{filter_significant_results}
\alias{filter_significant_results}
\title{Filter Significant Limma Results and Add ROC Metrics}
\usage{
filter_significant_results(
  limma_results,
  p_val = 0.05,
  FC_cut = 1.1,
  add_ROC = TRUE
)
}
\arguments{
\item{limma_results}{Output from run_limma_analysis()}

\item{p_val}{Numeric; p-value threshold}

\item{FC_cut}{Numeric; fold change threshold}

\item{add_ROC}{Logical; calculate ROC metrics?}
}
\value{
List containing:
\item{sig_results}{Significant results with ROC}
\item{data_sig}{Significant expression matrix}
\item{data_sig_RC}{Row-centered significant data}
\item{metadata_sig}{Metadata for significant samples}
}
\description{
Filter Significant Limma Results and Add ROC Metrics
}
