% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_functions.R
\name{generate_annotation_colors}
\alias{generate_annotation_colors}
\title{Generate Annotation Colors for Heatmap}
\usage{
generate_annotation_colors(metadata)
}
\arguments{
\item{metadata}{Data frame with annotation variables}
}
\value{
Named list of color vectors for each annotation variable
}
\description{
Creates a named list of color palettes for heatmap annotations
}
\note{
Version 1.0.0
File: limma_function.R
}
\keyword{internal}
