% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pn_limma.R
\name{generate_heatmap_data}
\alias{generate_heatmap_data}
\title{Generate Heatmap Data}
\usage{
generate_heatmap_data(
  data_sig,
  data_sig_RC,
  meta_2,
  variable,
  continuous,
  class1,
  class2,
  add_anno
)
}
\arguments{
\item{data_sig}{Matrix of significant features}

\item{data_sig_RC}{Matrix of significant features (row-centered)}

\item{meta_2}{Metadata data frame}

\item{variable}{Character, primary variable name}

\item{continuous}{Logical, is variable continuous?}

\item{class1}{Character, class 1 name (categorical only)}

\item{class2}{Character, class 2 name (categorical only)}

\item{add_anno}{Character vector, additional annotation columns}
}
\value{
List with heatmap data components
}
\description{
Generate Heatmap Data
}
\keyword{internal}
