% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_func_dev_functions.R
\name{generate_limma_plots}
\alias{generate_limma_plots}
\title{Generate All Limma Plots}
\usage{
generate_limma_plots(
  limma_results,
  sig_results,
  p_val = 0.05,
  FC_cut = 1.1,
  plot_width = 20,
  plot_height = 10,
  add_anno = NULL,
  plot_violins = TRUE
)
}
\arguments{
\item{limma_results}{Output from run_limma_analysis()}

\item{sig_results}{Output from filter_significant_results()}

\item{p_val}{Numeric; p-value threshold}

\item{FC_cut}{Numeric; fold change threshold}

\item{plot_width}{Numeric; plot width}

\item{plot_height}{Numeric; plot height}

\item{add_anno}{Character vector; additional annotation columns}

\item{plot_violins}{Logical; generate violin plots?}
}
\value{
List of plot objects and heatmap data
}
\description{
Generate All Limma Plots
}
