% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pn_limma.R
\name{generate_roc_results}
\alias{generate_roc_results}
\title{Generate ROC Results}
\usage{
generate_roc_results(
  exprs_data,
  data_sig,
  meta_2,
  TT,
  variable,
  class1,
  class2,
  sig_rows,
  fc_direction = "both"
)
}
\arguments{
\item{exprs_data}{Matrix, full expression data}

\item{data_sig}{Matrix, significant features only}

\item{meta_2}{Metadata data frame}

\item{TT}{TopTable results}

\item{variable}{Character, primary variable name}

\item{class1}{Character, class 1 name}

\item{class2}{Character, class 2 name}

\item{sig_rows}{Integer vector, indices of significant rows}

\item{fc_direction}{Character, "up", "down", or "both" (default "both")}
}
\value{
Data frame with merged limma + ROC results
}
\description{
Generate ROC Results
}
\keyword{internal}
