% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pn_limma.R
\name{generate_violin_plots}
\alias{generate_violin_plots}
\title{Generate Violin Plots}
\usage{
generate_violin_plots(exprs_data, merged_results, meta_2, variable)
}
\arguments{
\item{exprs_data}{Matrix, full expression data}

\item{merged_results}{Data frame, merged limma + ROC results OR topTable subset}

\item{meta_2}{Metadata data frame}

\item{variable}{Character, primary variable name}
}
\value{
List of ggplot objects
}
\description{
Generate Violin Plots
}
\keyword{internal}
