% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_func_dev_functions.R
\name{generate_volcano_plot}
\alias{generate_volcano_plot}
\title{Generate Volcano Plot}
\usage{
generate_volcano_plot(topTable, sig_features, p_val, FC_cut)
}
\description{
Generate Volcano Plot
}
\keyword{internal}
