% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_utils.R
\name{get_clinical_samples}
\alias{get_clinical_samples}
\title{Get Clinical Sample IDs}
\usage{
get_clinical_samples(eset, PN_column = "Sample_type", PN_value = "TR")
}
\arguments{
\item{eset}{ExpressionSet object}

\item{PN_column}{Character; column name identifying PN samples}

\item{PN_value}{Character; value identifying PN samples}
}
\value{
Character vector of clinical sample IDs
}
\description{
Extract sample IDs for clinical (non-PN) samples from ExpressionSet
}
