% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_utils.R
\name{get_expression_data}
\alias{get_expression_data}
\title{Extract Expression Data from ExpressionSet}
\usage{
get_expression_data(eset, assay_name = NULL)
}
\arguments{
\item{eset}{ExpressionSet object}

\item{assay_name}{Character; name of assay (default = NULL uses exprs())}
}
\value{
Matrix of expression data
}
\description{
Utility function to extract expression matrix from ExpressionSet
}
