% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{get_expset_assay_names}
\alias{get_expset_assay_names}
\title{Extract ExpressionSet Names and AssayData Elements}
\usage{
get_expset_assay_names(ExpSet_list)
}
\arguments{
\item{ExpSet_list}{A list of ExpressionSet objects}
}
\value{
A named list where names are ExpressionSet names and values are vectors of assayData element names
}
\description{
Creates a named list of all assayData elements for each ExpressionSet in the list
}
\examples{
\dontrun{
ExpSet_names <- get_expset_assay_names(ExpSet_list)
}
}
