% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_functions.R
\name{get_ExpSet_name}
\alias{get_ExpSet_name}
\title{Get ExpressionSet Name from Assay Name}
\usage{
get_ExpSet_name(name, ExpSet_list)
}
\arguments{
\item{name}{Character; assay name (e.g., "clinical_loess_normalised")}

\item{ExpSet_list}{List; list of ExpressionSets}
}
\value{
Character; name of ExpressionSet containing the assay
}
\description{
Helper function to determine which ExpressionSet contains the requested assay
}
\keyword{internal}
