% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_utils.R
\name{get_PN_options}
\alias{get_PN_options}
\title{Get Available PN Options from Metadata}
\usage{
get_PN_options(eset)
}
\arguments{
\item{eset}{ExpressionSet object}
}
\value{
Named list of column -> unique values
}
\description{
Return all possible PN identifier options from metadata
}
