% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_helpers.R
\name{get_template_path}
\alias{get_template_path}
\title{Get Path to Template File}
\usage{
get_template_path(template)
}
\arguments{
\item{template}{Character; template name}
}
\value{
Path to template file
}
\description{
Get Path to Template File
}
\examples{
get_template_path("denoiser_pipeline")
}
