% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_functions.R
\name{heatmap_module_ui}
\alias{heatmap_module_ui}
\title{Heatmap UI Module}
\usage{
heatmap_module_ui(id)
}
\arguments{
\item{id}{A unique string identifier for the module namespace.}
}
\value{
A \code{uiOutput()} element containing the heatmap plot.
}
\description{
This UI module renders a heatmap output container for the corresponding server-side module.
}
\note{
Version 1.0 from
heatmap_functions.R
}
