% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_functions.R
\docType{data}
\name{KL_palette}
\alias{KL_palette}
\title{KL Color Palette}
\format{
A character vector of hex color codes.
}
\usage{
KL_palette
}
\description{
A qualitative color palette compiled from all available qualitative palettes in \code{RColorBrewer}.
Designed to provide a large number of distinct colors for categorical annotations (e.g., heatmaps).
}
\note{
Version 1.0 from
colour_functions.R
}
\keyword{datasets}
