% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_ShinyModule.R
\name{limmaHeatmapUI}
\alias{limmaHeatmapUI}
\title{Limma Heatmap UI Module}
\usage{
limmaHeatmapUI(id)
}
\arguments{
\item{id}{Character string, namespace ID for the module}
}
\value{
Shiny UI elements
}
\description{
UI component for interactive heatmap of significant features
}
\note{
Version 1.0.0
File: limma_ShinyModule.R
}
