% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_ShinyModule.R
\name{limmaResultsUI}
\alias{limmaResultsUI}
\title{Limma Results Summary UI Module}
\usage{
limmaResultsUI(id)
}
\arguments{
\item{id}{Character string, namespace ID for the module}
}
\value{
Shiny UI elements
}
\description{
UI component for displaying limma results summary and tables
}
\note{
Version 1.0.0
File: limma_ShinyModule.R
}
