% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_ShinyModule.R
\name{limmaViolinServer}
\alias{limmaViolinServer}
\title{Limma Violin Plot Server Module}
\usage{
limmaViolinServer(id, limma_results)
}
\arguments{
\item{id}{Character string, namespace ID matching the UI}

\item{limma_results}{Reactive expression returning limma analysis results}
}
\value{
NULL (generates plot output)
}
\description{
Server logic for generating violin plots
}
\note{
Version 1.0.0
File: limma_ShinyModule.R
}
