% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_ShinyModule.R
\name{limmaViolinUI}
\alias{limmaViolinUI}
\title{Limma Violin Plot UI Module}
\usage{
limmaViolinUI(id)
}
\arguments{
\item{id}{Character string, namespace ID for the module}
}
\value{
Shiny UI elements
}
\description{
UI component for violin plots of significant features
}
\note{
Version 1.0.0
File: limma_ShinyModule.R
}
