% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_ShinyModule.R
\name{limmaVolcanoUI}
\alias{limmaVolcanoUI}
\title{Limma Volcano Plot UI Module}
\usage{
limmaVolcanoUI(id)
}
\arguments{
\item{id}{Character string, namespace ID for the module}
}
\value{
Shiny UI elements
}
\description{
UI component for interactive volcano plot of limma results
}
\note{
Version 1.0.0
File: limma_ShinyModule.R
}
