% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_helpers.R
\name{list_templates}
\alias{list_templates}
\title{List Available Pipeline Templates}
\usage{
list_templates()
}
\value{
Character vector of available template names
}
\description{
List Available Pipeline Templates
}
\examples{
list_templates()
}
