% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_functions.R
\name{logratio2foldchange}
\alias{logratio2foldchange}
\title{Convert Log Ratio to Fold Change}
\usage{
logratio2foldchange(logFC)
}
\arguments{
\item{logFC}{Numeric vector of log2 fold change values}
}
\value{
Numeric vector of fold change values
}
\description{
Converts log2 fold change values to linear fold change
}
\note{
Version 1.0.0
File: limma_function.R
}
\examples{
logratio2foldchange(c(1, -1, 2, -2))
}
