% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_Function.R
\name{Merge_QC_function}
\alias{Merge_QC_function}
\title{Merge QC Function}
\usage{
Merge_QC_function(QC, type = "PerSample", feature_column = "Protein")
}
\arguments{
\item{QC}{A nested list of QC results}

\item{type}{Either "PerSample" or "PerProtein", determines the axis of aggregation}

\item{feature_column}{Character string for the feature identifier (default = "Protein")}
}
\value{
A list with:
\itemize{
\item flagged_df: full QC results with metrics
\item flagged: vector of flagged Sample or feature_column values
\item Avg_df: thresholds and average values per QC metric
}
}
\description{
Aggregates QC data from a structured QC object across either samples or features.
}
\examples{
Merge_QC_function(QC_list, type = "PerSample")

}
\author{
Shaun Garnett
}
