% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudule_functions.R
\name{mid}
\alias{mid}
\title{Create Mode-Specific Module ID}
\usage{
mid(base_id, mode_suffix = NULL)
}
\arguments{
\item{base_id}{Character string. The base module ID.}

\item{mode_suffix}{Character string. Optional suffix.  If NULL, searches
for MODE_SUFFIX in calling environment.  Returns base_id if not found.}
}
\value{
Character string. Module ID with suffix (if available).
}
\description{
Create Mode-Specific Module ID
}
