% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_annotation_analysis.R
\name{mod_annotation_analysis_server}
\alias{mod_annotation_analysis_server}
\title{Annotation Analysis Module - Server}
\usage{
mod_annotation_analysis_server(id, eset, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet}

\item{debug}{Enable debug mode}
}
\description{
Annotation Analysis Module - Server
}
