% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_annotation_analysis.R
\name{mod_annotation_analysis_ui}
\alias{mod_annotation_analysis_ui}
\title{Annotation Analysis Module - UI}
\usage{
mod_annotation_analysis_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug button}
}
\description{
Analyzes phenoData columns to identify useful grouping variables
}
