% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manifest_upload.R
\name{mod_annotation_distribution_server}
\alias{mod_annotation_distribution_server}
\title{Annotation Distribution Testing Module - Server}
\usage{
mod_annotation_distribution_server(
  id,
  manifest_data,
  batch_columns,
  annotation_columns,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{manifest_data}{Reactive data frame from manifest upload}

\item{batch_columns}{Reactive character vector - Internal batch column names}

\item{annotation_columns}{Reactive character vector - Client annotation column names}
}
\description{
Annotation Distribution Testing Module - Server
}
