% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app_data_selection.R
\name{mod_app_data_selection_server}
\alias{mod_app_data_selection_server}
\title{App Data Selection Module - Reusable Server}
\usage{
mod_app_data_selection_server(id, default_selection = NULL, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{default_selection}{Default ExpSet to select}

\item{debug}{Enable debug mode}
}
\description{
App Data Selection Module - Reusable Server
}
