% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app_data_selection.R
\name{mod_app_data_selection_ui}
\alias{mod_app_data_selection_ui}
\title{App Data Selection Module - Reusable UI}
\usage{
mod_app_data_selection_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug controls}
}
\description{
App Data Selection Module - Reusable UI
}
