% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch_column_selector.R
\name{mod_batch_column_selector_ui}
\alias{mod_batch_column_selector_ui}
\title{Batch Testing Column Selector - UI}
\usage{
mod_batch_column_selector_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug button}
}
\description{
Select columns for batch effect testing
}
