% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch_combined_analysis.R
\name{mod_batch_combined_analysis_server}
\alias{mod_batch_combined_analysis_server}
\title{Combined Batch Analysis Module - Server}
\usage{
mod_batch_combined_analysis_server(
  id,
  eset,
  sample_group_column,
  batch_columns,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet}

\item{sample_group_column}{Reactive character Sample grouping column name}

\item{batch_columns}{Reactive character vector Batch columns to test}

\item{debug}{Enable debug mode}
}
\description{
Combined Batch Analysis Module - Server
}
