% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch.R
\name{mod_batch_distribution_test_server}
\alias{mod_batch_distribution_test_server}
\title{Batch Distribution Testing Module - Server}
\usage{
mod_batch_distribution_test_server(
  id,
  eset,
  sample_group_column,
  batch_columns,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet}

\item{sample_group_column}{Reactive character.   Sample grouping column name}

\item{batch_columns}{Reactive character vector.  Batch column names to test}

\item{debug}{Enable debug mode}
}
\description{
Batch Distribution Testing Module - Server
}
