% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch_distribution_test.R
\name{mod_batch_distribution_test_ui}
\alias{mod_batch_distribution_test_ui}
\title{Batch Distribution Testing Module - UI}
\usage{
mod_batch_distribution_test_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug button}
}
\description{
Tests if sample groups are evenly distributed across batch columns
}
