% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch.R
\name{mod_batch_testing_server}
\alias{mod_batch_testing_server}
\title{Batch Testing Module - Server}
\usage{
mod_batch_testing_server(id, eset, selected_columns, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet}

\item{selected_columns}{Reactive vector of column names to test}

\item{debug}{Enable debug mode}
}
\description{
Batch Testing Module - Server
}
