% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch_testing.R
\name{mod_batch_testing_ui}
\alias{mod_batch_testing_ui}
\title{Batch Testing Module - UI}
\usage{
mod_batch_testing_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug button}
}
\description{
Run ANOVA tests on selected batch columns
}
