% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch_visualization.R
\name{mod_batch_visualization_server}
\alias{mod_batch_visualization_server}
\title{Batch Effect Visualization Module - Server}
\usage{
mod_batch_visualization_server(
  id,
  eset_original_name = reactive(NULL),
  eset_original,
  eset_corrected = reactive(NULL),
  sample_group_column = reactive(NULL),
  batch_factors = reactive(NULL),
  ExpSet_list = NULL,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset_original}{Reactive ExpressionSet (original data)}

\item{eset_corrected}{Reactive ExpressionSet (ComBat corrected data)}

\item{debug}{Enable debug mode}
}
\description{
Batch Effect Visualization Module - Server
}
