% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch_visualization.R
\name{mod_batch_visualization_ui}
\alias{mod_batch_visualization_ui}
\title{Batch Effect Visualization Module - UI}
\usage{
mod_batch_visualization_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug button}
}
\description{
Compare batch effects before and after correction using multiple visualization methods
}
