% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_column.R
\name{mod_column_selector_ui}
\alias{mod_column_selector_ui}
\title{Generic Column Selector - UI}
\usage{
mod_column_selector_ui(
  id,
  label = "Select Column",
  help_text = NULL,
  show_summary = TRUE,
  show_info = TRUE,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{label}{Label for the selector (e. g., "Sample Grouping Column", "Batch Testing Columns")}

\item{help_text}{Optional help text}

\item{show_summary}{Show summary box (default TRUE)}

\item{show_info}{Show info bubble (default TRUE)}

\item{debug}{Show debug button}
}
\description{
Reusable column selector for phenoData
}
