% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_combat_correction.R
\name{mod_combat_correction_selector_server}
\alias{mod_combat_correction_selector_server}
\title{ComBat Correction Selector Module - Server}
\usage{
mod_combat_correction_selector_server(
  id,
  eset,
  combined_results = reactive(NULL),
  debug = FALSE
)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{eset}{Reactive ExpressionSet.}

\item{combined_results}{Reactive data frame from combined batch analysis.}

\item{debug}{Logical. Enable debug mode (default FALSE).}
}
\description{
ComBat Correction Selector Module - Server
}
