% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_combat_correction.R
\name{mod_combat_correction_selector_ui}
\alias{mod_combat_correction_selector_ui}
\title{ComBat Correction Selector Module - UI}
\usage{
mod_combat_correction_selector_ui(id, show_info = TRUE, debug = FALSE)
}
\arguments{
\item{id}{Character.  Module namespace ID.}

\item{show_info}{Logical. Show info bubble (default TRUE).}

\item{debug}{Logical. Show debug button (default FALSE).}
}
\description{
Minimal dropdown for batch factor selection with expandable settings
}
