% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_combat_correction.R
\name{mod_combat_correction_server}
\alias{mod_combat_correction_server}
\title{ComBat Batch Correction Module - Server}
\usage{
mod_combat_correction_server(
  id,
  eset,
  sample_group_column,
  combined_results,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet}

\item{sample_group_column}{Reactive character.  Sample grouping column (to preserve)}

\item{combined_results}{Reactive data.frame. Results from combined batch analysis}

\item{debug}{Enable debug mode}
}
\description{
ComBat Batch Correction Module - Server
}
