% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_combat_correction.R
\name{mod_combat_correction_ui}
\alias{mod_combat_correction_ui}
\title{ComBat Batch Correction Module - UI}
\usage{
mod_combat_correction_ui(id, show_auto_run_toggle = TRUE, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug button}
}
\description{
Select batch factors and run ComBat correction
}
