% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_combat_correction.R
\name{mod_combat_multi_assay_server}
\alias{mod_combat_multi_assay_server}
\title{Multi-Assay Batch Correction Server}
\usage{
mod_combat_multi_assay_server(
  id,
  ExpSet_list,
  update_ExpSet_list,
  sample_group_column,
  selector,
  default_target_assays = NULL,
  debug = FALSE
)
}
\arguments{
\item{id}{Module ID}

\item{ExpSet_list}{Reactive containing list of ExpressionSets}

\item{update_ExpSet_list}{Function to update the ExpSet_list}

\item{selector}{Reactive list with combat settings (batch_factors, combat_model, par_prior, strategy, sample_group)}

\item{debug}{Logical, enable debug mode}
}
\description{
Multi-Assay Batch Correction Server
}
