% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_combat_correction.R
\name{mod_combat_multi_assay_ui}
\alias{mod_combat_multi_assay_ui}
\title{Multi-Assay Batch Correction UI}
\usage{
mod_combat_multi_assay_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module ID}
}
\description{
Multi-Assay Batch Correction UI
}
