% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_combat_correction.R
\name{mod_combat_single_server}
\alias{mod_combat_single_server}
\title{Single ComBat Correction Module - Server}
\usage{
mod_combat_single_server(
  id,
  eset,
  sample_group_column,
  combined_results = reactive(NULL),
  selector = NULL,
  show_auto_run_toggle = TRUE,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet}

\item{sample_group_column}{Reactive sample group column name}

\item{combined_results}{Reactive batch analysis results}

\item{selector}{ComBat selector module return object}

\item{show_auto_run_toggle}{Logical. Enable auto-run toggle}

\item{debug}{Enable debug mode}
}
\description{
Single ComBat Correction Module - Server
}
