% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_combat_correction.R
\name{mod_combat_single_ui}
\alias{mod_combat_single_ui}
\title{Single ComBat Correction Module - UI}
\usage{
mod_combat_single_ui(id, show_auto_run_toggle = TRUE, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{show_auto_run_toggle}{Logical.  Show auto-run toggle switch (default TRUE).}

\item{debug}{Show debug button}
}
\description{
Displays correction status and results with expandable advanced details
}
