% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_ShinyModule.R
\name{mod_denoiser_server}
\alias{mod_denoiser_server}
\title{Denoiser Module Server}
\usage{
mod_denoiser_server(id, ExpSet_list, eset_raw, eset_norm = NULL)
}
\arguments{
\item{id}{Character; module namespace ID}

\item{eset_raw}{Reactive; ExpressionSet with raw/NetI data}

\item{eset_norm}{Reactive; ExpressionSet with normalized data (optional)}
}
\description{
Denoiser Module Server
}
