% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_selector_ShinyModule.R
\name{mod_eset_selector_server}
\alias{mod_eset_selector_server}
\title{ExpressionSet Selector Module Server}
\usage{
mod_eset_selector_server(
  id,
  ExpSet_list,
  default_selection = "clinical_loess_normalised"
)
}
\arguments{
\item{id}{Character; module namespace ID}

\item{ExpSet_list}{Reactive; list of ExpressionSets}

\item{default_selection}{Character; default selection if available (e.g., "clinical_loess_normalised")}
}
\value{
Reactive ExpressionSet with exprs() set to selected assay
}
\description{
ExpressionSet Selector Module Server
}
