% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eset.R
\name{mod_eset_selector_standalone_server}
\alias{mod_eset_selector_standalone_server}
\title{Standalone ExpressionSet Selector Server with Subset & Transform
Complete data preparation pipeline}
\usage{
mod_eset_selector_standalone_server(
  id,
  ExpSet_list,
  ExpSet_list_version = reactive(0),
  default_selection = reactive(NULL),
  enable_subset = TRUE,
  enable_transform = TRUE,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{ExpSet_list}{Reactive returning named list of ExpressionSets}

\item{default_selection}{Default ExpSet name to select}

\item{enable_subset}{Enable subsetting (default TRUE)}

\item{enable_transform}{Enable transformation (default TRUE)}

\item{debug}{Enable debug mode (default FALSE)}
}
\description{
Standalone ExpressionSet Selector Server with Subset & Transform
Complete data preparation pipeline
}
