% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpSet_selector_ShinyModule.R
\name{mod_eset_selector_ui}
\alias{mod_eset_selector_ui}
\title{ExpressionSet Selector Module UI}
\usage{
mod_eset_selector_ui(id)
}
\arguments{
\item{id}{Character; module namespace ID}
}
\description{
ExpressionSet Selector Module UI
}
