% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eset.R
\name{mod_eset_subset_server}
\alias{mod_eset_subset_server}
\title{ExpressionSet Subsetting Module - Server}
\usage{
mod_eset_subset_server(id, eset, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet (original)}

\item{debug}{Enable debug mode}
}
\description{
ExpressionSet Subsetting Module - Server
}
