% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eset_subset.R
\name{mod_eset_subset_ui}
\alias{mod_eset_subset_ui}
\title{ExpressionSet Subsetting Module - UI}
\usage{
mod_eset_subset_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug button}
}
\description{
Subset ExpressionSet by pData or fData columns
}
