% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eset.R
\name{mod_eset_transform_server}
\alias{mod_eset_transform_server}
\title{ExpressionSet Transformation Module - Server}
\usage{
mod_eset_transform_server(id, eset, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet (input)}

\item{debug}{Enable debug mode}
}
\description{
ExpressionSet Transformation Module - Server
}
