% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_expset_export.R
\name{mod_expset_export_server}
\alias{mod_expset_export_server}
\title{ExpressionSet List Export Module - Server}
\usage{
mod_expset_export_server(id, ExpSet_list)
}
\arguments{
\item{id}{Module namespace ID}

\item{ExpSet_list}{Reactive returning the ExpressionSet list}
}
\description{
ExpressionSet List Export Module - Server
}
