% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_expset_export.R
\name{mod_expset_export_ui}
\alias{mod_expset_export_ui}
\title{ExpressionSet List Export Module - UI}
\usage{
mod_expset_export_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Export/save the entire ExpressionSet list
}
