% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_expset_manager.R
\name{mod_expset_manager_ui}
\alias{mod_expset_manager_ui}
\title{ExpressionSet Manager Module - UI}
\usage{
mod_expset_manager_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Manage and export ExpressionSet list with ComBat corrections
}
