% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_expset_metadata_manager.R
\name{mod_expset_metadata_manager_server}
\alias{mod_expset_metadata_manager_server}
\title{ExpressionSet Metadata Manager - Server}
\usage{
mod_expset_metadata_manager_server(
  id,
  ExpSet_list,
  update_ExpSet_list = NULL,
  master_eset_name = "RawData_ExpSet",
  debug = FALSE
)
}
\arguments{
\item{id}{Character string. Namespace identifier.}

\item{ExpSet_list}{Reactive. List of ExpressionSets.}

\item{update_ExpSet_list}{Function. Callback to update ExpSet_list in parent.}

\item{master_eset_name}{Character. Name of ExpSet with master metadata (default "RawData_ExpSet").}

\item{debug}{Logical. Enable debug mode (default FALSE).}
}
\value{
Reactive list with updated ExpSet_list
}
\description{
ExpressionSet Metadata Manager - Server
}
