% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_expset_metadata_manager.R
\name{mod_expset_metadata_manager_ui}
\alias{mod_expset_metadata_manager_ui}
\title{ExpressionSet Metadata Manager - UI}
\usage{
mod_expset_metadata_manager_ui(id)
}
\arguments{
\item{id}{Character string. Namespace identifier.}
}
\description{
ExpressionSet Metadata Manager - UI
}
