% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_expset_viewer.R
\name{mod_expset_viewer_server}
\alias{mod_expset_viewer_server}
\title{ExpressionSet Viewer Module - Server}
\usage{
mod_expset_viewer_server(id, ExpSet_list)
}
\arguments{
\item{id}{Module namespace ID}

\item{ExpSet_list}{Reactive returning named list of ExpressionSets}
}
\description{
ExpressionSet Viewer Module - Server
}
