% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manifest_upload.R
\name{mod_manifest_upload_ui}
\alias{mod_manifest_upload_ui}
\title{Annotation Distribution Testing Module - UI}
\usage{
mod_manifest_upload_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Tests if client annotations are evenly distributed across internal batch columns
}
