% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pc_removal_visualizer.R
\name{mod_pc_visualizer_server}
\alias{mod_pc_visualizer_server}
\title{PC Removal Visualizer Module Server}
\usage{
mod_pc_visualizer_server(id, eset_raw, denoiser_results)
}
\arguments{
\item{id}{Module ID}

\item{eset_raw}{Reactive expression returning raw ExpressionSet}

\item{denoiser_results}{Reactive expression returning results from denoiser module}
}
\description{
PC Removal Visualizer Module Server
}
