% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pc_removal_visualizer.R
\name{mod_pc_visualizer_ui}
\alias{mod_pc_visualizer_ui}
\title{PC Removal Visualizer Module UI}
\usage{
mod_pc_visualizer_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
PC Removal Visualizer Module UI
}
