% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pn_limma.R
\name{mod_pn_limma_ui}
\alias{mod_pn_limma_ui}
\title{PN Limma Analysis Module - UI}
\usage{
mod_pn_limma_ui(id)
}
\arguments{
\item{id}{Character string. Namespace identifier.}
}
\description{
PN Limma Analysis Module - UI
}
