% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pn_visualisation.R
\name{mod_pn_viz_ui}
\alias{mod_pn_viz_ui}
\title{Pooled Normal Visualization Module UI}
\usage{
mod_pn_viz_ui(
  id,
  title = "Data Visualization",
  status = "info",
  show_density = TRUE
)
}
\arguments{
\item{id}{Character; module namespace ID}

\item{title}{Character; title for the box}

\item{status}{Character; box status color (default "info")}

\item{show_density}{Logical; whether to show density plot (default TRUE)}
}
\description{
Pooled Normal Visualization Module UI
}
