% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_group_selector.R
\name{mod_sample_group_selector_server}
\alias{mod_sample_group_selector_server}
\title{Sample Group Column Selector - Server}
\usage{
mod_sample_group_selector_server(
  id,
  eset,
  default_column = "Labels",
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet}

\item{default_column}{Character.  Default column name (e.g., "Labels")}

\item{debug}{Enable debug mode}
}
\description{
Sample Group Column Selector - Server
}
