% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_group_selector.R
\name{mod_sample_group_selector_ui}
\alias{mod_sample_group_selector_ui}
\title{Sample Group Column Selector - UI}
\usage{
mod_sample_group_selector_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug button}
}
\description{
Select primary sample grouping column from phenoData
}
