% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCF_functions.R
\name{NCF_Corr}
\alias{NCF_Corr}
\title{NCF Correlation Calculation Function}
\usage{
NCF_Corr(input_data, zz = "ZZ_con2")
}
\arguments{
\item{input_data}{A numeric matrix or data frame of features (rows = features, columns = samples).}

\item{zz}{Character string specifying the name of the reference feature (default is \code{"ZZ_con2"}).}
}
\value{
A list containing:
\item{df_ZZ}{A \code{data.table} with feature names and their correlation to the \code{zz} reference.}
\item{q_ZZ}{Quantiles of the correlation values.}
\item{zz}{The name of the reference feature used.}
}
\description{
Computes Pearson correlation between all features in the input matrix and a specified ZZ reference feature (e.g., \code{"ZZ_con2"}),
returning a data table of correlation values and their quantiles.
}
\note{
Version 1.0 from
NCF_function.R
}
