% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pn_limma.R
\name{performance_metrics}
\alias{performance_metrics}
\title{Calculate Performance Metrics for AutoAbs}
\usage{
performance_metrics(input, metadata, variable, flush, groupPos, groupNeg)
}
\arguments{
\item{input}{Matrix, expression data}

\item{metadata}{Data frame, sample metadata}

\item{variable}{Character, grouping variable}

\item{flush}{Data frame, limma results with frequency data}

\item{groupPos}{Character, positive group name}

\item{groupNeg}{Character, negative group name}
}
\value{
Data frame with performance metrics
}
\description{
Calculate Performance Metrics for AutoAbs
}
\keyword{internal}
