% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFC_ShinyModule.R
\name{pFC_Server}
\alias{pFC_Server}
\title{pFC Analysis Shiny Module - Server}
\usage{
pFC_Server(id, eset_reactive, default_var_reactive = reactive(NULL))
}
\arguments{
\item{id}{Module namespace ID}

\item{eset_reactive}{Reactive expression returning an ExpressionSet object}

\item{default_var_reactive}{Reactive expression returning the default variable column name}
}
\description{
Server component for pFC analysis in Shiny apps
}
