% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFC_ShinyModule.R
\name{pFC_UI}
\alias{pFC_UI}
\title{pFC Analysis Shiny Module - UI}
\usage{
pFC_UI(id, use_box = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{use_box}{Logical; wrap in shinydashboard box? (default: FALSE)}
}
\description{
UI component for pFC parameter selection in Shiny apps
}
