% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_visualization_plots.R
\name{plot_correlation_histogram}
\alias{plot_correlation_histogram}
\title{Plot Correlation Histogram}
\usage{
plot_correlation_histogram(correlation_results)
}
\description{
Plot Correlation Histogram
}
