% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_visualization_plots.R
\name{plot_correlation_intra_inter}
\alias{plot_correlation_intra_inter}
\title{Plot Intra vs Inter Batch Correlation}
\usage{
plot_correlation_intra_inter(
  correlation_results,
  meta,
  batch_column,
  tr_samples
)
}
\description{
Plot Intra vs Inter Batch Correlation
}
