% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pc_removal_visualizer.R
\name{plot_density_data}
\alias{plot_density_data}
\title{Plot density data}
\usage{
plot_density_data(
  expr_data,
  eset,
  group_var = "Labels",
  plot_type = "overlay",
  title = "test"
)
}
\description{
Plot density data
}
\keyword{internal}
