% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pc_removal_visualizer.R
\name{plot_difference_data}
\alias{plot_difference_data}
\title{Plot difference data}
\usage{
plot_difference_data(diff_data, eset, group_var)
}
\description{
Plot difference data
}
\keyword{internal}
