% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_visualization_plots.R
\name{plot_distance_heatmap}
\alias{plot_distance_heatmap}
\title{Plot Distance Heatmap}
\usage{
plot_distance_heatmap(dist_mat, meta, color_by, title = "")
}
\description{
Plot Distance Heatmap
}
