% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pc_removal_visualizer.R
\name{plot_heatmap_data}
\alias{plot_heatmap_data}
\title{Plot heatmap data}
\usage{
plot_heatmap_data(
  expr_data,
  eset,
  group_var = "Labels",
  scale_type = "none",
  show_features = TRUE,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  title = "test"
)
}
\description{
Plot heatmap data
}
\keyword{internal}
