% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pc_removal_visualizer.R
\name{plot_loadings_heatmap}
\alias{plot_loadings_heatmap}
\title{Plot loadings heatmap}
\usage{
plot_loadings_heatmap(loadings, sdev, top_n)
}
\description{
Plot loadings heatmap
}
\keyword{internal}
