% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_visualization_plots.R
\name{plot_pca_variance_comparison}
\alias{plot_pca_variance_comparison}
\title{Plot PCA Variance Comparison}
\usage{
plot_pca_variance_comparison(var_orig, var_corr)
}
\description{
Plot PCA Variance Comparison
}
