% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_visualization_plots.R
\name{plot_pca_variance_single}
\alias{plot_pca_variance_single}
\title{Plot PCA Variance (single dataset)}
\usage{
plot_pca_variance_single(variance)
}
\description{
Plot PCA Variance (single dataset)
}
