% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pc_removal_visualizer.R
\name{plot_removed_density_facet}
\alias{plot_removed_density_facet}
\title{Plot removed components density - faceted by component}
\usage{
plot_removed_density_facet(removed_components_list, eset, group_var = "Labels")
}
\description{
Plot removed components density - faceted by component
}
\keyword{internal}
