% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_visualization_plots.R
\name{plot_sample_correlation_intra_inter}
\alias{plot_sample_correlation_intra_inter}
\title{Plot Sample Correlation Intra vs Inter Batch}
\usage{
plot_sample_correlation_intra_inter(
  correlation_results,
  meta,
  batch_column,
  samples
)
}
\description{
Plot Sample Correlation Intra vs Inter Batch
}
