% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pc_removal_visualizer.R
\name{plot_sample_statistics}
\alias{plot_sample_statistics}
\title{Plot sample statistics}
\usage{
plot_sample_statistics(expr_data, eset, group_var, title)
}
\description{
Plot sample statistics
}
\keyword{internal}
