% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_visualization_plots.R
\name{plot_tsne_comparison}
\alias{plot_tsne_comparison}
\title{Plot t-SNE Comparison (side-by-side)}
\usage{
plot_tsne_comparison(tsne_orig, tsne_corr, meta, color_by, shape_by = NULL)
}
\description{
Plot t-SNE Comparison (side-by-side)
}
