% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholds_Protein_1.6.R
\name{proteinsServer}
\alias{proteinsServer}
\title{Protein Threshold Server Module}
\usage{
proteinsServer(id, global_preset, product_positions)
}
\arguments{
\item{id}{Module namespace ID}

\item{global_preset}{Reactive value containing preset name}

\item{product_positions}{Reactive value containing product position data}
}
\value{
Reactive value containing product data
}
\description{
Protein Threshold Server Module
}
