% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholds_Protein_1.6.R
\name{proteinsUI}
\alias{proteinsUI}
\title{Protein Threshold UI Module}
\usage{
proteinsUI(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Protein Threshold UI Module
}
