% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSA_type_functions.R
\docType{data}
\name{PSA_ROs_pkg}
\alias{PSA_ROs_pkg}
\title{PSA_ROs_pkg gene list}
\format{
A character vector of gene names.
}
\usage{
data(PSA_ROs_pkg)
}
\description{
A vector of prostate-specific antigens used for quality control.
}
\note{
Version 1.0 from
tools/PSA_ROs_data.R
}
\keyword{datasets}
