% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_Function.R
\name{QC_ctrl_repot_plots}
\alias{QC_ctrl_repot_plots}
\title{QC Control Report Plots}
\usage{
QC_ctrl_repot_plots(
  datCollate,
  data_type = "PerSample",
  feature_column = "Protein"
)
}
\arguments{
\item{datCollate}{A list containing input data and parameters.}

\item{data_type}{Character. Either "PerSample" or "PerProtein". Default is "PerSample".}

\item{feature_column}{Character. The name of the feature column (e.g., "Protein"). Default is "Protein".}
}
\value{
A list of QC outputs including Cy3 control CVs and IgG dilution R² results.
}
\description{
Generates QC metrics for control probes (e.g., BSA and IgG), including CV and R² checks.
Allows flexibility in feature column naming (e.g., "Protein" or "Feature").
}
\note{
Version 1.0.0 from
QC_functions.R
}
