% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots.R
\name{QC_line_plot_list_function}
\alias{QC_line_plot_list_function}
\title{QC Line Plot List Function}
\usage{
QC_line_plot_list_function(
  Data,
  data,
  QC,
  data_name,
  grouping_column,
  x,
  feature_column = "Protein"
)
}
\arguments{
\item{Data}{Raw or filtered input data frame (typically probe-level).}

\item{data}{Full QC object list (with nested \code{[[QC]][[data_name]]$data} structure).}

\item{QC}{Character. Name of the QC list entry (e.g., 'ctrl_QC').}

\item{data_name}{Character. Name of the sub-data entry (e.g., 'Cy3AvgCV').}

\item{grouping_column}{Character. Column for grouping traces (e.g., 'Sample' or 'Feature').}

\item{x}{Character. Column name to use for the x-axis.}

\item{feature_column}{Character. Column name of features to clean (default = "Protein").}
}
\value{
A \code{plotly} object created via \code{QC_line_plotly_function()}.
}
\description{
Prepares and formats data for line-based QC plots using provided grouping.
}
\note{
Version 1.0.0 from
QC_sub_functions.R
}
