% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_utils.R
\name{quick_inspect_eset}
\alias{quick_inspect_eset}
\title{Quick Inspect ExpressionSet for Denoising}
\usage{
quick_inspect_eset(eset, assay_name = NULL)
}
\arguments{
\item{eset}{ExpressionSet object}

\item{assay_name}{Character; assay to inspect}
}
\description{
Print detailed diagnostics about an ExpressionSet
}
