% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_func_dev_functions.R
\name{save_limma_results}
\alias{save_limma_results}
\title{Save Limma Results to Files}
\usage{
save_limma_results(
  limma_results,
  sig_results = NULL,
  plots = NULL,
  file_descriptor,
  p_val = 0.05,
  plot_width = 20,
  plot_height = 10
)
}
\arguments{
\item{limma_results}{Output from run_limma_analysis()}

\item{sig_results}{Output from filter_significant_results()}

\item{plots}{Output from generate_limma_plots()}

\item{file_descriptor}{Character; output folder name}

\item{p_val}{Numeric; p-value threshold}

\item{plot_width}{Numeric; plot width}

\item{plot_height}{Numeric; plot height}
}
\description{
Save Limma Results to Files
}
