% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholds_Spot_1.5.R
\name{spotsServer}
\alias{spotsServer}
\title{Spot Threshold Server Module}
\usage{
spotsServer(id, global_preset, product_positions)
}
\arguments{
\item{id}{Module namespace ID}

\item{global_preset}{Reactive value containing preset name}

\item{product_positions}{Reactive value containing product position data}
}
\value{
Reactive value containing product data
}
\description{
Spot Threshold Server Module
}
