% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholds_Spot_1.5.R
\name{spotsUI}
\alias{spotsUI}
\title{Spot Threshold UI Module}
\usage{
spotsUI(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Spot Threshold UI Module
}
