% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_helpers.R
\name{view_template}
\alias{view_template}
\title{View Template in Editor}
\usage{
view_template(template)
}
\arguments{
\item{template}{Character; template name}
}
\description{
View Template in Editor
}
\examples{
\dontrun{
view_template("denoiser_pipeline")
}
}
