% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch_column_selector.R
\name{mod_batch_column_selector_server}
\alias{mod_batch_column_selector_server}
\title{Batch Testing Column Selector - Server}
\usage{
mod_batch_column_selector_server(
  id,
  eset,
  useful_columns,
  default_columns = NULL,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{eset}{Reactive ExpressionSet}

\item{useful_columns}{Reactive vector of useful column names from annotation analysis}

\item{default_columns}{Reactive or static vector of default selections}

\item{debug}{Enable debug mode}
}
\description{
Batch Testing Column Selector - Server
}
