% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_batch.R
\name{mod_batch_combined_analysis_ui}
\alias{mod_batch_combined_analysis_ui}
\title{Combined Batch Analysis Module - UI}
\usage{
mod_batch_combined_analysis_ui(id, show_auto_run_toggle = TRUE, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{show_auto_run_toggle}{Logical. Show auto-run toggle switch (default TRUE).}

\item{debug}{Show debug button}
}
\description{
Combined ANOVA batch effects and distribution confounding tests
}
