% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eset.R
\name{mod_eset_transform_ui}
\alias{mod_eset_transform_ui}
\title{ExpressionSet Transformation Module - UI}
\usage{
mod_eset_transform_ui(id, debug = FALSE)
}
\arguments{
\item{id}{Module namespace ID}

\item{debug}{Show debug button}
}
\description{
Transform expression data (log, unlog, row-scale)
}
