% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_expset_viewer.R
\name{mod_expset_viewer_server}
\alias{mod_expset_viewer_server}
\title{ExpressionSet Viewer Module - Server}
\usage{
mod_expset_viewer_server(
  id,
  ExpSet_list,
  default_selection = NULL,
  enable_subset = TRUE,
  enable_transform = TRUE,
  debug = FALSE
)
}
\arguments{
\item{id}{Module namespace ID}

\item{ExpSet_list}{Reactive returning named list of ExpressionSets}

\item{default_selection}{Default ExpSet to select (optional)}

\item{enable_subset}{Enable subsetting (default TRUE)}

\item{enable_transform}{Enable transformation (default TRUE)}

\item{debug}{Enable debug mode (default FALSE)}
}
\description{
ExpressionSet Viewer Module - Server
}
