% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_expset_viewer.R
\name{mod_expset_viewer_ui}
\alias{mod_expset_viewer_ui}
\title{ExpressionSet Viewer Module - UI}
\usage{
mod_expset_viewer_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Comprehensive viewer for ExpressionSet objects with download functionality
}
