% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pn_limma.R
\name{mod_pn_limma_server}
\alias{mod_pn_limma_server}
\title{Enhanced Limma Analysis Module - Server}
\usage{
mod_pn_limma_server(id, eset, default_assay = "exprs")
}
\arguments{
\item{id}{Character string. Namespace identifier.}

\item{eset}{Reactive or static. ExpressionSet object.}

\item{default_assay}{Character or reactive. Default assayData element (default "exprs").}
}
\value{
Reactive list containing limma results
}
\description{
Enhanced Limma Analysis Module - Server
}
