% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pn_limma.R
\name{mod_pn_limma_server}
\alias{mod_pn_limma_server}
\title{PN Limma Analysis Module - Server}
\usage{
mod_pn_limma_server(id, eset_raw, eset_norm = NULL)
}
\arguments{
\item{id}{Namespace ID}

\item{eset_raw}{Reactive returning raw ExpressionSet}

\item{eset_norm}{Reactive returning normalized ExpressionSet (optional)}
}
\value{
Reactive list containing:
\itemize{
\item exp_PN_AAbs: numeric vector of expected range
\item PN_AAbs: character vector of specific antigens
\item limma_results: full limma analysis results object
}
}
\description{
PN Limma Analysis Module - Server
}
