% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{mode_function}
\alias{mode_function}
\title{Estimate the mode of a numeric vector using kernel density}
\usage{
mode_function(values)
}
\arguments{
\item{values}{A numeric vector.}
}
\value{
A single numeric value representing the mode, or \code{NA} if not enough data.
}
\description{
Estimate the mode of a numeric vector using kernel density
}
\note{
File:maths_function.R

This function estimates the mode (most frequent value) of a numeric vector
by calculating the peak of its kernel density estimate.
If fewer than 4 non-NA values are present, it returns NA.
}
\examples{
mode_function(c(1, 2, 2, 3, 4))
mode_function(c(NA, 2, 2, NA, 3))
}
